requirements2.txt
africastalking==1.2.4,amqp==1.4.9
arabic-reshaper==2.1.3
arviz==0.11.4
asgiref==3.4.1
bcrypt==3.2.0
billiard==3.3.0.23
cachetools==4.2.4
celery==3.1.26.post2
certifi==2021.10.8
cffi==1.15.0
cftime==1.5.1.1
charset-normalizer==2.0.8
config==0.5.1
contextlib2==21.6.0
cycler==0.11.0
decouple==0.0.7
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
Django==3.2.9
django-celery==3.3.1
django-email-verification==0.1.0
django-filter==21.1
django-import-export==2.6.1
et-xmlfile==1.1.0
fastprogress==1.0.0
filelock==3.4.0
fonttools==4.28.3
future==0.18.2
html5lib==1.1
idna==3.3
kiwisolver==1.3.2
kombu==3.0.37
MarkupPy==1.14
matplotlib==3.5.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
netCDF4==1.5.8
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
patsy==0.5.2
Pillow==8.4.0
protobuf==3.19.1
pycparser==2.21
pymc3==3.11.4
pyparsing==3.0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
reportlab==3.6.3
requests==2.26.0
schema==0.7.4
scipy==1.7.3
semver==2.13.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
Theano-PyMC==1.1.2
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
xarray==0.20.2
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
django-bootstrap-v5